+++ src/app/np-directives-demo/np-directives-demo.module.ts
@@ -3,8 +3,7 @@ import { CommonModule } from '@angular/common';
 
 import { NpDirectivesDemoRoutingModule } from './np-directives-demo-routing.module';
 import { NpDirectivesDemoComponent } from './np-directives-demo.component';
-import { NpCardModule, NpAutofocusModule, NpHightlightModule, NpOrderByModule } from 'np-ui-lib';
-import { NpDeferModule } from 'projects/np-ui-lib/src/public-api';
+import { NpCardModule, NpAutofocusModule, NpHightlightModule, NpOrderByModule, NpDeferModule } from 'np-ui-lib';
 
 @NgModule({
   declarations: [NpDirectivesDemoComponent],